rsync new code

+++ crystallib/installers/web/mdbook/mdbook.v
@@ -26,8 +26,7 @@ pub fn install(args_ InstallArgs) ! {
 		args.reset = true
 	}
 
-	for plname in ['mdbook-mermaid', 'mdbook-echarts', 'mdbook-kroki-preprocessor', 'mdbook-pdf',
-		'mdbook-last-changed'] {
+	for plname in ['mdbook-mermaid', 'mdbook-echarts', 'mdbook-kroki-preprocessor'] {
 		if !osal.cmd_exists(plname) {
 			console.print_header('did not find: ${plname}')
 			args.reset = true
@@ -50,6 +49,7 @@ pub fn build() ! {
 	if osal.is_linux() {
 		dest_on_os = '/usr/local/bin'
 	}	
+	mut ok:=false
 	cmd := '
 	set +ex
 	rm ${os.home_dir()}/.cargo/bin/mdb* 2>&1 >/dev/null
@@ -58,21 +58,27 @@ pub fn build() ! {
 	source ~/.cargo/env
 	cargo install mdbook
 	cargo install mdbook-mermaid
-	cargo install mdbook-last-changed
 	cargo install mdbook-echarts
+	cargo install mdbook-kroki-preprocessor
+
+	#cargo install mdbook-last-changed
 	#cargo install mdbook-embed
 	#cargo install mdbook-plantuml
-	cargo install mdbook-kroki-preprocessor
-	cargo install mdbook-pdf --features fetch
+	#cargo install mdbook-pdf --features fetch
 	#cargo install mdbook-linkcheck
-	cp ${os.home_dir()}.cargo/bin/mdb* ${dest_on_os}/	
+
+	cp ${os.home_dir()}/.cargo/bin/mdb* ${dest_on_os}/	
 	'
 	defer {		
-		println("ERROR IN INSTALL MDBOOK, WILL ABORT")
-		osal.execute_stdout("rm ${os.home_dir()}/.cargo/bin/mdb*") or {}
-		osal.execute_stdout("rm ${dest_on_os}/mdb*") or {}
+		if ! ok{
+			println("ERROR IN INSTALL MDBOOK, WILL ABORT")
+			osal.execute_stdout("rm ${os.home_dir()}/.cargo/bin/mdb*") or {}
+			osal.execute_stdout("rm ${dest_on_os}/mdb*") or {}
+		}
 	}
 	osal.execute_stdout(cmd)!
+	ok=true
 	osal.done_set('install_mdbook', 'OK')!
 	console.print_header('mdbook installed')
+
 }
